refactor(stortbook): remove duplicated code
